+++ tests/Console/ModelsCommand/Comment/__snapshots__/Test__test__1.php
@@ -24,7 +24,7 @@ use Illuminate\Database\Eloquent\Relations\MorphTo;
  * @property-read \Illuminate\Database\Eloquent\Collection<int, Simple> $relationHasMany HasMany relations.
  * @property-read int|null $relation_has_many_count
  * @property-read Simple|null $relationHasOne Others relations.
- * @property-read Model|\Eloquent $relationMorphTo MorphTo relations.
+ * @property-read Model|null $relationMorphTo MorphTo relations.
  * @property-write mixed $first_name Set the user's first name.
  * @method static \Illuminate\Database\Eloquent\Builder|Simple active() Scope a query to only include active users.
  * @method static \Illuminate\Database\Eloquent\Builder|Simple newModelQuery()