+++ CHANGELOG.md
@@ -16,6 +16,7 @@ All notable changes to this project will be documented in this file.
 - Casts without properties are now removed
 - `belongsTo` relationships pointing to a column that is not a foreign key will not always be marked as nullable anymore
 - Relationships pointing to a model that has global scopes will now always be marked as nullable.
+- `morphTo` relationship type is now always `Model|null`
 
 ### Added
 - Add support for intersection types